+++ src/main/java/com/ats/adminpanel/model/logistics/ServDetailAddPart.java
@@ -14,8 +14,8 @@ public class ServDetailAddPart {
 	private String vehName;
 	private int sprId; 
 	private String partName;
-	private int sprQty; 
-	private int sprRate; 
+	private float sprQty; 
+	private float sprRate; 
 	private float sprTaxableAmt; 
 	private float sprTaxAmt;   
 	private float sprTaxAmtPer;
@@ -74,16 +74,16 @@ public class ServDetailAddPart {
 	public void setPartName(String partName) {
 		this.partName = partName;
 	}
-	public int getSprQty() {
+	public float getSprQty() {
 		return sprQty;
 	}
-	public void setSprQty(int sprQty) {
+	public void setSprQty(float sprQty) {
 		this.sprQty = sprQty;
 	}
-	public int getSprRate() {
+	public float getSprRate() {
 		return sprRate;
 	}
-	public void setSprRate(int sprRate) {
+	public void setSprRate(float sprRate) {
 		this.sprRate = sprRate;
 	}
 	public float getSprTaxableAmt() {
